mhd: cleanup pass over the dual-order surface
Seven independent tidy-ups with no behaviour change beyond the two noted:

- MHD4 now refuses a reconstruction it cannot honour. ReconstructionSelector
  silently accepted Constant/Linear/WENO3 at fourth order; PHARE_Types asserts
  the same WENOZ-or-MP5 rule pharein already enforces in check_mhd_profile.
- nbrGhostsFromReconstruction is templated on MHDOpts::MHDOrder rather than a
  raw uint32_t, so the order cannot be built out of an arbitrary integer, and
  MHDOptions keeps the enum instead of casting it away.
- PointValueHandler -> PointValueConverter: it converts, and the name now says
  so. point_value_handler_utils.hpp, which only ever held PointValueState, is
  renamed to match its contents.
- The twelve face/edge-to-cell-centre stencils collapse to six templated on
  Order. They stay constexpr, not consteval: they are passed by address as
  template arguments, and taking the address of a consteval function is
  ill-formed.
- MC2011TemporalTransfer::assembledStateNames() returned a reference into a
  mutable optional it re-emplaced on every call, invalidating any reference a
  caller still held. The names are built once in registerQuantities(), where
  they are already known.
- EulerUsingComputedFlux stamped the moment timestamps onto `state` while the
  B line two rows below stamped `statenew`, which is what
  fillConservativeGhosts then reads. The asymmetry was the bug.
- Comments citing internal derivation notes are replaced by the substance they
  pointed at. Published-paper citations are untouched.

+++ src/core/models/options/mhd_options.hpp
@@ -23,7 +23,7 @@ struct MHDFieldOptions
 
     auto static constexpr reconstruction_nghosts
         = MHDOpts::reconstruction_nghosts_v<opts.reconstruction_type>;
-    auto static constexpr mhd_order = static_cast<std::uint32_t>(opts.mhd_order);
+    auto static constexpr mhd_order = opts.mhd_order;
 
     auto static constexpr field_ghost_width
         = core::nbrGhostsFromReconstruction<reconstruction_nghosts, mhd_order>();
